requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
delta-rest-client>=1.0.13
streamlit>=1.25.0
plotly>=5.0.0
matplotlib>=3.5.0
python-dotenv>=0.21.0
requests>=2.28.0
